+++ backend/canisters/group_index/impl/src/queries/mod.rs
@@ -1,5 +1,6 @@
 pub mod active_groups;
 pub mod explore_communities;
+pub mod explore_groups;
 pub mod filter_groups;
 pub mod http_request;
 pub mod recommended_groups;