+++ backend/canisters/user/impl/src/crypto.rs
@@ -1,5 +1,5 @@
 use crate::read_state;
-use ledger_utils::{nns, sns};
+use ledger_utils::{icrc1, nns, sns};
 use types::{CompletedCryptoTransaction, FailedCryptoTransaction, PendingCryptoTransaction, UserId};
 
 pub async fn process_transaction(
@@ -37,5 +37,6 @@ async fn process_transaction_internal(
     match transaction {
         PendingCryptoTransaction::NNS(t) => nns::process_transaction(t, my_user_id.into(), ledger_canister_id).await,
         PendingCryptoTransaction::SNS(t) => sns::process_transaction(t, my_user_id.into(), ledger_canister_id).await,
+        PendingCryptoTransaction::ICRC1(t) => icrc1::process_transaction(t, my_user_id.into(), ledger_canister_id).await,
     }
 }