+++ backend/libraries/ledger_utils/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod icrc1;
 pub mod nns;
 pub mod sns;
 
@@ -35,6 +36,12 @@ pub fn calculate_transaction_hash(sender: CanisterId, args: &TransferArgs) -> Tr
     };
 
     transaction.hash()
+}
+
+pub fn format_crypto_amount(units: u128, decimals: u32) -> String {
+    let subdividable_by = 10u128.pow(decimals);
+
+    format!("{}.{:0}", units / subdividable_by, units % subdividable_by)
 }
 
 /// An operation which modifies account balances