+++ backend/libraries/types/src/group_activity.rs
@@ -6,8 +6,8 @@ pub struct PublicCommunityActivity {
     pub timestamp: TimestampMillis,
     pub member_count: u32,
     pub channel_count: u32,
-    // pub last_hour: Activity,
-    // pub last_day: Activity,
+    pub last_hour: Activity,
+    pub last_day: Activity,
 }
 
 #[derive(Serialize, Deserialize, Debug, Default)]
@@ -21,8 +21,6 @@ pub struct PublicGroupActivity {
 
 #[derive(Serialize, Deserialize, Debug, Default)]
 pub struct Activity {
-    #[serde(alias = "participant_count_change")]
-    pub member_count_change: i32,
     pub messages: u32,
     pub message_unique_users: u32,
     pub reactions: u32,